chore: Remove unused reference types in next-env.d.ts and making app router specific fetches
